+++ @navikt/core/react/src/form/search/search.stories.tsx
@@ -1,5 +1,6 @@
-import { Meta, StoryObj } from "@storybook/react-vite";
+import { Meta } from "@storybook/react-vite";
 import React, { useState } from "react";
+import { renderStoriesForChromatic } from "../../util/renderStoriesForChromatic";
 import Search from "./Search";
 
 export default {
@@ -33,8 +34,6 @@ export default {
     chromatic: { disable: true },
   },
 } satisfies Meta<typeof Search>;
-
-type Story = StoryObj<typeof Search>;
 
 export const Default = (props) => {
   const [state, setState] = useState("");
@@ -175,56 +174,16 @@ export const ColorRole = () => (
   </div>
 );
 
-export const Chromatic: Story = {
-  render: (...props) => (
-    <div>
-      <div>
-        <h2>Default</h2>
-        <Default {...props} />
-      </div>
-      <div>
-        <h2>Small</h2>
-        <Small />
-      </div>
-      <div>
-        <h2>Variants</h2>
-        <Variants />
-      </div>
-      <div>
-        <h2>ErrorVariants</h2>
-        <ErrorVariants />
-      </div>
-      <div>
-        <h2>Disabled</h2>
-        <Disabled />
-      </div>
-      <div>
-        <h2>Placeholder</h2>
-        <Placeholder />
-      </div>
-      <div>
-        <h2>Text</h2>
-        <Text />
-      </div>
-      <div>
-        <h2>WLabel</h2>
-        <WLabel />
-      </div>
-      <div>
-        <h2>NoClearButton</h2>
-        <NoClearButton />
-      </div>
-      <div>
-        <h2>HtmlSize</h2>
-        <HtmlSize />
-      </div>
-      <div>
-        <h2>ColorRole</h2>
-        <ColorRole />
-      </div>
-    </div>
-  ),
-  parameters: {
-    chromatic: { disable: false },
-  },
-};
+export const Chromatic = renderStoriesForChromatic({
+  Default,
+  Small,
+  Variants,
+  ErrorVariants,
+  Disabled,
+  Placeholder,
+  Text,
+  WLabel,
+  NoClearButton,
+  HtmlSize,
+  ColorRole,
+});